test authentication endpoints

+++ app/graphql/types/item_type.rb
@@ -3,7 +3,7 @@ module Types
         field :id, ID, null: false
         field :product, Types::ProductType, null: false
         field :quantity, Integer, null: false
-        field :unit_price, Decimal, null: false
+        field :unit_price, Float, null: false
         field :created_at, GraphQL::Types::ISO8601DateTime, null: false
         field :updated_at, GraphQL::Types::ISO8601DateTime, null: false
     end
